Stop LWIP UDP checksum failures
Flakiness in the tapcp SNAP comms was observed and traced to the tge_rx module.
Data were received OK from the MAC, written into the dual-port BRAM seemingly
OK (verified by ILA). But, when read by the CPU side of the buffer, they were
corrupt.
Repeated reads of the packets from RAM yielded the same corrupt packets every time,
suggesting the problem was on the write side of the interface, even though the write
side looks fine via ILA.

ILAs indicated no funny business with read/write address clashes. When a corrupted
word was received, it was not the previous RAM value, or any other systematically wrong
data.
When the mac clock and wishbone clock are set to the same source, this problem goes away.

The code highlights (hat-tip to the author) that the MSB of the read-side of the RAM
interface is on the wrong clock domain. However, the state machine logic would seem
to prevent the signal from doing anything untoward during a read. ILAs confirmed this to
be the case.
However, registering this bit to the correct (i.e. CPU) clock domain makes the UDP checksum
problem go away. It is not clear (to me) why this would be the case, unless the mismatched
address bit clock domain was causing some tool-related issue.
Prior to this fix, examining the timing for the signal from the RAM's MSB (on the mac clk)
to the RAM pins didn't show anything suspicious.

This commit double registers the address MSB into the cpu clock domain prior to use on
the CPU side of the RAM. No analysis was done to figure out if this extra latency could present
an issue. However, this seems unlikely -- similar registering, and then a CPU-side poll, is
used to trigger the CPU to read data.

This code was tested on a SNAP board using a loop of writes to sys_scratchpad with random values.
After 1M writes (at ~1500 writes per second) no errors were observed. I.e., there were no
TFTP retries.

Microblaze version was:
```
JAM: built 2020/04/30-13:24+0000 jackh@rtr-dev1
Flags -DUSE_SPI -DUSE_ICAP -DUSE_XADC
Version v1.0.0-553-gfb57c29581
```

This commit also changes some register attributes from shreg_extract=no to async_reg=yes,
to correctly indicate that they are clock-domain-crossing registers. This change was tested
by itself, and did not fix the data corruption issue

TODO: Regenerate the SNAP golden image with this fix.