Add RS-FEC support
100GbE now has two versions of the CMAC IP -- one with RSFEC and
one without. Use the rsfec keywork in the platform yaml file
to decide which to use.

Currently, if the platform enables FEC, it is always switched on.
TODO: on a platform which supports FEC, expose ports / registers
to disable it, so that the user can use 100G standards which
don't implement it.