Respect manually-specified iostd
The pin loc/iostd parsing is a bit of a mess -- the PinConstraint
parsing respects the user's spec, and writes it in to the castro yaml file,
but then the castro parser was overriding the loc/iostd based on the
symbolic name stored in the yaml. Whose job is it to turn symbolic
constraints into physical ones?