Add ability to pass block-specific parameters from YAML
Often it is useful to be able to make some platform-specific
alterations to yellow blocks. This commit has the YellowBlock
class look for a yaml key in a platform's config with a name
matching the yellow block class name. The contents of this
key are then available to the yellowblock under the
attribute `blockconf`. Hopefully this is neater and more general
adding lots of `if self.platform.name == ...` clauses into
yallow block methods