Make RFDC ports Unsigned
Possibly signed makes more sense if one sample per clock is being presented to the FPGA.
But, it definitely doesn't make sense if there is more than one sample in a bus,
and I think therefore the interface should be consistently unsigned.
TODO: Maybe a switch which makes the interface signed iff the user is only
using one sample per port would be the best solution.