Remove PFB init's reaching out into adder tree
The PFB initialization attempts to smartly minimize the number of bits in the
PFB add trees, but figuring out what scaling the coefficients will yield.
However, the add tree has it's own init function, and sets the scale blocks
to a value which assumes the add tree is operating at full precision.

This commit lets the add tree init win this initialization battle, at the expense
of a lower-than necessary gain through the PFB. But... no model initialization errors!