Comply with new RFSoC 100GbE requirement
I don't know why a new 100GbE class has been created for the RFSoC
(the previous block was being used on an RFSoC HTG ZRF16 without issue?)
but the new class seemingly requires a YAML entry indicating which control
pins to use.
Probably using None, as this commit implements, isn't the right thing to do,
but it is at least consistent with behaviour using the old core.

TODO:
  1. put the right thing in the YAML file to control ZRF16 QSFP control lines
  2. Consider refactoring the two current 100G classes into one, or at least
     re-using more of the code so there isn't complete duplication for the
     sake of a few differences.