Display 4 DACs even for DT parts
When using two DACs of a DT tile, these should be dac0 and dac1.
Looking at the IPI RFD block, it seems that dac1 ends up generating the port
for the second upconverter path of the tile.
This commit displays all 4 dacs for DT parts. The user should select
dac0 and dac2, else it won't work, but at least this gets through the compile.

There is undoubtedly a better way to deal with this.