Add PFB_FIR support for oversample by 2x, 4x
In the 2x case, the block rearranges output samples so they can be fed
straight into an FFT. (This is not new behaviour, but the 2x case
used to be enabled with a checkbox).
In the 4x case, the block outputs the 4 blocks of time samples. It is
assumed the user will feed these into some downsampling arrangement to get
(eg) 4/3 oversampling.
The FIR block should probably be able to arrange samples appropriately,
but it isn't entirely clear what format is best, since this depends what the user
is going to do. We could make some assumptions like: oversampling by 4x is only
used when downsampling by >2x, in which case half the output samples could be ignored.
TODO: Figure out what to do with this downstream logic and absorb appropriate parts into FIR