Support DPSS window without MATLAB signal processing toolbox
Use a MATLAB function call, but....

scipy seems to work when called within matlab, but the
dpss window function fails with an error relating to eigenvalue
decomposition. This seems to be related to LAPACK drivers
and my suspicion is that MATLAB is using its own, which
aren't playing nicely with scipy.

So, create a complete hack around where the scipy window function is launched
using the system's python.
This will fail is the system doesn't have a 'python' command (eg because it
only has 'python3') or if the system python doesn't have scipy
installed